make connection tokens stateless
