+++ services/system/Accounts/plugin/src/db/keys.rs
@@ -3,6 +3,5 @@ pub struct DbKeys;
 impl DbKeys {
     pub const LOGGED_IN: &'static str = "logged-in";
     pub const CONNECTED_ACCOUNTS: &'static str = "connected-accounts";
-    pub const CONNECT_TOKENS: &'static str = "connect-tokens";
     pub const CONNECTED_APPS: &'static str = "connected-apps";
 }
