+++ services/system/Accounts/plugin/src/db/mod.rs
@@ -1,8 +1,2 @@
 pub mod apps_table;
-pub use apps_table::*;
-
-pub mod tokens_table;
-pub use tokens_table::*;
-
 pub mod keys;
-pub use keys::*;
