+++ services/system/Accounts/plugin/src/interfaces/admin.rs
@@ -3,8 +3,8 @@ use crate::plugin::AccountsPlugin;
 use crate::bindings::exports::accounts::plugin::admin::{AppDetails, Guest as Admin};
 use crate::bindings::exports::accounts::plugin::api::Guest as API;
 use crate::bindings::host::common::client as Client;
-
-use crate::db::*;
+use crate::connection_token::*;
+use crate::db::apps_table::*;
 use crate::helpers::*;
 
 // Asserts that the caller is the active app, and that it's the `accounts` app.
@@ -31,14 +31,10 @@ impl Admin for AccountsPlugin {
     fn decode_connection_token(token: String) -> Option<AppDetails> {
         assert_caller_admin();
 
-        if let Some(token) = TokensTable::new().decode(&token) {
-            Some(AppDetails {
-                origin: token.app_origin,
-                app: token.app,
-            })
-        } else {
-            None
-        }
+        ConnectionToken::from_str(&token).map(|t| AppDetails {
+            origin: t.app_origin,
+            app: t.app,
+        })
     }
 
     fn get_connected_apps() -> Vec<String> {
